solve issue

+++ qustion5.c
@@ -5,14 +5,14 @@
 
 
 int main(){
-    int n,sq;
+    float n,sq;
     printf("Enter number: ");
     scanf("%d",&n);
 
     // printf("Enter exp: ");
     // scanf("%d",&m);
     sq = pow(n,2);
-    printf("The squre of %d is %d",n,sq);
+    printf("The squre of %d is %f",n,sq);
 
     return 0;
 }
